Add unity activation / return license workflow
Add 2 new tasks to the unity plugin convention

`Activate` will try to use the credentials provided in unity plugin
extension to activate unity, If a license is provided it tries to
activate unity with license.

`ReturnLicense` task will return the assigned license if available.

This feature works only if the credentials are set in unity plugin
extension or provided via gradle properties or environment values

It is possible to switch off `autoActivation` and `autoReturnLicense`
within the unity plugin extension.