Fix logCategory default value
Description
===========

With the last changes we somehow brought in a regression for the
`logCategory` value. It seems we no longer evaluate clojure values at
one point. Since we try to limit API's with clojures that is Ok. I run
the tests and the one testing the `logCategory` also failed (strange).
So I changed the default mapping to the old `conventionMapping` style
and it works.

Changes

![FIX] `logCategory` default convention mapping