Fix compatibility for gradle > 4.4
Description
===========

We used some internal API's that changed in gradle 4.4.
The API `resolveLater` in `PathToFileResolver` is gone so we replicate
this API with `new Factory<File>` invocations instead for now.
The default constructor for `DefaultExecHandleBuilder` changed in a way
that backwards compatibility is no longer possible. I added a delegate
to a new field `execHandleBuilder` which gets created with a
`DefaultExecActionFactory`

Changes
=======

![FIX] usage of removed API `resolveLater`
![CHANGE] internal `BatchModeAction` exec handle builder logic
![CHANGE] use `FileResolver` instead of `PathToFileResolver` in code