Fix unity 2019 logfile handling (#52)
Description
===========

Unity 2019 changed the behavior of the `-logFile` switch. To print to
`stdout` we must provide `-` as the file path which signals Unity to
print the log to `stdout`. This behavior works also on windows now so I
enabled the logoutput on windows when the Unity version is >= 2019.

Most of the test setup is very brittle at the moment. The tests don't
run with a real Unity installation because of setup and runtime
considerations. The way I path the unity version for tests is not
optimal and will change in the future.

To execute the version check for all batchmode actions, I added a
`protected` getter to fetch the Unity version beeing executed. I still
use the `ArtifactVersion` object but I have plans to change this in the
future and make this part of the public API.

Changes
=======

![FIX] ![UNITY] 2019 logfile handling
![ADD] [unity-version-manager-jni] as dependency to fetch the Unity version

[unity-version-manager-jni]:    https://github.com/wooga/unity-version-manager-jni