Fix unity activation (#64)
## Description

It turns out the unity activation routine will open and import a project. Since we don't provide a project path for the `activate` and `returnLicense` tasks this can lead to build errors because Unity will pick the last project opened on the system. This picked project could be incompatible and fail the activation task.

This is a very simple patch to provide the project path to `activate` and `returnLicense` tasks. The whole structure should be cleaned up in another change. The separation of  AbstractUnityProjectTask` and `AbstractUnityActivationTask` makes no real sense anymore.

## Changes

* ![FIX] ![UNITY] activation task