Enable batchmode for tests in Unity >= 2018 (#72)
Description
===========

Unity `5.6` added a complete new test runner compared to the one used in
`5.5`. The main problem was a bug in unity. When the tests where
executed in _batchmode_ unity actually never executed anything. This was
true also in `2017` versions of unity. We tested this with `2018` and it
looks like the issue was fixed. I enable _batchmode_ for all versions >=
2018.

Changes
=======

* ![IMPROVE] enable _batchmode_ for tests in unity >= 2018