Fix getBatchModeForEditModeTest and getBatchModeForPlayModeTest in extension
Description
===========

I made a small error in the getter implemention for both
`getBatchModeForEditModeTest` and `getBatchModeForPlayModeTest` in the
`DefaultUnityPluginExtension` class. I used a `Boolean` class to
make sure that the default value would be `null` not `false`.
My problem was that the `if` check I provided also checks if the
value is `false`. In this case the body was not executed and the
default value was being calculated from properties/environment
which falls back to `true`. I want to return the value present
in the to be checked `Boolean` field if it is not `null` so
I made the check more explicit.

Changes
=======

* ![FIX] `getBatchModeForEditModeTest` implementation
* ![FIX] `getBatchModeForPlayModeTest` implementation