Fix remove default BuildTarget.undefined
Description
===========

The old plugin version only allowed to set the BuildTarget flag
with an enum. To check if a value is provided, the enum contained
a value `undefined` so the task impl could check if the flag is actually
provided. This is no longer needed and adds a small bug when the
flag is not set. I removed the enum value and all mentions in code and adjusted
the tests.

Changes
=======

* ![FIX] remove default `BuildTarget.undefined`