added unity root, unity dotnet and unity mono framework properties (#115)
Created properties for key executables/folders inside unity folder structure. As unity folder structure differs by OS, we try first to find the root folder by name, and if this fails, we just go up the file tree a bunch of times hoping to hit the right directory. Then, from this root directory, we try to find the dotnet executable and mono framework folder. All of these can be overwritten by user by informing them explicitly on the extension.

All of these file tree-related operations are on the UnityFileTree class.

Co-authored-by: Joaquim Neto <joaquim.neto@wooga.net>